+++ packages/apps/loan-process/src/module/Calculator.Slider.js
@@ -1,46 +1,68 @@
 import React from "react"
-import { Slider2, CurrencyInput, Layout } from "@staccx/base"
+import { Slider2, CurrencyInput, Layout, theming } from "@staccx/base"
 
 class CalculatorSlider extends React.Component {
   state = {
     sum: parseInt(this.props.defaultValue, 10),
-    sliderSum: parseInt(this.props.defaultValue, 10)
+    sliderSum: parseInt(this.props.defaultValue, 10),
+    step: this.props.step,
+    percentage: parseInt(this.props.defaultValue, 10) / this.props.max
   }
 
   onInputChange = value => {
     if (value) {
-      this.setState({ sum: parseInt(value, 10) })
-      if (this.state.sum > this.props.min && this.state.sum < this.props.max) {
-        this.setState({ sliderSum: parseInt(value, 10) })
+      const sum = parseInt(value, 10)
+      this.setState({
+        sum,
+        percentage: sum / this.props.max
+      })
+      this.props.onChange(sum)
+      if (
+        this.state.sum >= this.props.min &&
+        this.state.sum <= this.props.max
+      ) {
+        this.setState({ sliderSum: sum })
       }
     } else {
-      this.setState({ sum: 0 })
+      this.setState({ sum: "" })
     }
   }
 
   onSliderChange = value => {
-    this.setState({ sum: parseInt(value, 10) })
-    this.setState({ sliderSum: parseInt(value, 10) })
+    const sum = parseInt(value, 10)
+    this.setState({
+      sum,
+      sliderSum: sum,
+      percentage: sum / this.props.max
+    })
+
+    this.props.onChange(sum)
   }
 
   render() {
-    const { name, label, max, min, step } = this.props
+    const { name, label, max, min } = this.props
 
     return (
       <div>
         <Layout rowGap="small">
           <CurrencyInput
             name={name}
+            id={name}
             label={label}
             value={this.state.sum}
             onChange={e => this.onInputChange(e.target.rawValue)}
+            onFocus={() => this.setState({ step: 1 })}
+            variant="calculatorSlider"
+            percentage={this.state.percentage}
           />
           <Slider2
             min={min}
             max={max}
-            step={1000}
+            step={this.state.step}
             defaultValue={this.state.sliderSum}
-            onChange={value => this.onSliderChange(value)}
+            onUpdate={value => this.onSliderChange(value)}
+            onFocus={() => this.setState({ step: this.props.step })}
+            variant={["calculatorSlider", theming.VARIANT_DEFAULT]}
           />
         </Layout>
       </div>