+++ packages/themes/Fundu/src/Theme/FunduTheme.js
@@ -12,6 +12,7 @@ import {
 import { TextStyling } from "./theme.Texts"
 import { ImageStyling } from "./theme.Images"
 import { ButtonStyling } from "./theme.Buttons"
+import { TableStyling } from "./theme.Tables"
 import {
   InputStyling,
   LabelStyling,
@@ -29,12 +30,15 @@ import {
   SelectSelectedWrapperStyling,
   PostalCodeInputStyling,
   SelectWrapperStyling,
-  FileInputStyles
+  FileInputStyles,
+  Slider2RailStyling,
+  Slider2TrackStyling
 } from "./theme.Inputs"
 import { LayoutStyling } from "./theme.Layouts"
 import { LayoutItemStyling } from "./theme.LayoutItem"
 import { LogoStyle } from "./Logo"
 import reset from "./Reset"
+import { hideVisually } from "polished"
 
 const borderRadius = "2px"
 const headerHeight = "70px"
@@ -74,14 +78,14 @@ const color = {
   white: "#fff",
   red: "#ea526f",
   pink: "#f8d0db",
-  green: "#00a190",
+  green: "#3ECF8E",
   blue: "#188EFB",
   yellow: "#f4da87",
   line: "#EBEBF2",
   wcag: "#777",
-  disabled: "#BABABA",
+  disabled: "#c1c1c1",
   warning: "#c54059",
-  black: "#000",
+  black: "#000000",
   subtleHover: "#F7F7FF",
   link: "#1F5D9B"
 }
@@ -120,7 +124,8 @@ export default new theming.Theme({
   targetSize,
   borderRadius,
   headerHeight,
-  fontFamily
+  fontFamily,
+  hideVisually
 })
   .add(reset)
   .add(HeadingStyling)
@@ -154,3 +159,6 @@ export default new theming.Theme({
   .add(PostalCodeInputStyling)
   .add(ImageStyling)
   .add(FileInputStyles)
+  .add(Slider2RailStyling)
+  .add(Slider2TrackStyling)
+  .add(TableStyling)