+++ packages/apps/loan-process/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.20.0"></a>
+# [8.20.0](https://bitbucket.org/stacc-flow/bento/compare/v8.19.0...v8.20.0) (2018-11-15)
+
+**Note:** Version bump only for package @staccx/loan-process
+
+
+
+
+
+<a name="8.19.0"></a>
+# [8.19.0](https://bitbucket.org/stacc-flow/bento/compare/v8.18.0...v8.19.0) (2018-11-14)
+
+
+### Features
+
+* **loan-process:** improve slider ([d07933e](https://bitbucket.org/stacc-flow/bento/commits/d07933e))
+* **loan-process:** style slider ([26be5c9](https://bitbucket.org/stacc-flow/bento/commits/26be5c9))
+* **loan-process:** use Table component for tables + style ([5209a37](https://bitbucket.org/stacc-flow/bento/commits/5209a37))
+
+
+
+
+
+<a name="8.18.0"></a>
+# [8.18.0](https://bitbucket.org/stacc-flow/bento/compare/v8.17.0...v8.18.0) (2018-11-13)
+
+
+### Features
+
+* **loan-process:** styling of fileupload Styrke + Fundu ([77ea74a](https://bitbucket.org/stacc-flow/bento/commits/77ea74a))
+
+
+
+
+
 <a name="8.17.0"></a>
 # [8.17.0](https://bitbucket.org/stacc-flow/bento/compare/v8.16.0...v8.17.0) (2018-11-13)