+++ packages/internal/mr-x/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.20.0"></a>
+# [8.20.0](https://bitbucket.org/stacc-flow/bento/compare/v8.19.0...v8.20.0) (2018-11-15)
+
+
+### Features
+
+* **mr-x:** parse text and link_names ([5f11484](https://bitbucket.org/stacc-flow/bento/commits/5f11484))
+
+
+
+
+
+<a name="8.19.0"></a>
+# [8.19.0](https://bitbucket.org/stacc-flow/bento/compare/v8.18.0...v8.19.0) (2018-11-14)
+
+
+### Features
+
+* **mr-x:** expose postMessage function ([e125329](https://bitbucket.org/stacc-flow/bento/commits/e125329))
+* **release-bento:** finished release script ([763c7b6](https://bitbucket.org/stacc-flow/bento/commits/763c7b6))
+
+
+
+
+
 <a name="8.15.0"></a>
 # [8.15.0](https://bitbucket.org/stacc-flow/bento/compare/v8.14.0...v8.15.0) (2018-11-12)